move BackupFile method to utils
